feat: add repository for bed
